refactor: 페이지 컴포넌트명을 더 구체적으로 변경
