+++ src/App.tsx
@@ -1,13 +1,15 @@
 import { Route, Routes } from 'react-router';
-import Complete from './pages/complete/Complete';
-import Payments from './pages/payments/Payments';
 import { ROUTE_PATH } from './constants/route';
+import { CardRegisterCompletePage, CardRegisterPage } from './pages';
 
 function App() {
   return (
     <Routes>
-      <Route path={ROUTE_PATH.home} element={<Payments />} />
-      <Route path={ROUTE_PATH.complete} element={<Complete />} />
+      <Route path={ROUTE_PATH.home} element={<CardRegisterPage />} />
+      <Route
+        path={ROUTE_PATH.complete}
+        element={<CardRegisterCompletePage />}
+      />
     </Routes>
   );
 }
