+++ src/pages/index.ts
@@ -1,0 +1,2 @@
+export { default as CardRegisterPage } from './CardRegisterPage';
+export { default as CardRegisterCompletePage } from './CardRegisterCompletePage';
